fix: rename component FiltersSidebar and update changelog
