+++ packages/core/docs/changelog/5684.js
@@ -1,6 +1,6 @@
 module.exports = {
     description: 'remove filters sidebar to separate component and add ignore to nuxt config',
-    link: '',
+    link: 'https://github.com/vuestorefront/vue-storefront/pull/5822',
     isBreaking: false,
     breakingChanges: [],
     author: 'Adam Pawliński',
