+++ packages/core/nuxt-theme-module/theme/components/FiltersSidebar.vue
